refactor(issues.show): choose better name for helper struct
The helper struct is called IssueItem and acts as a wrapper for Issue
objects, which also provides an `execute` function. Another api
function, like issues.delete, will also need a helper struct to provide
a _different_ `execute` function. To prevent naming collision and
confusion, the name for the helper struct(s) should be a little more
specific.